+++ build.rs
@@ -3,6 +3,7 @@ use std::env;
 
 fn main() {
     println!("cargo::rustc-check-cfg=cfg(ptrace_supported)");
+    println!("cargo::rustc-check-cfg=cfg(ptrace_default)");
     println!("cargo::rustc-check-cfg=cfg(tarpaulin_include)");
     println!("cargo::rustc-check-cfg=cfg(tarpaulin)");
     println!("cargo::rustc-check-cfg=cfg(nightly)");
@@ -18,7 +19,12 @@ fn main() {
 
     let target_arch = env::var("CARGO_CFG_TARGET_ARCH").expect("CARGO_CFG_TARGET_ARCH not set");
     let target_os = env::var("CARGO_CFG_TARGET_OS").expect("CARGO_CFG_TARGET_OS not set");
-    if target_os == "linux" && (target_arch == "x86_64" || target_arch == "x86") {
+    if target_os == "linux"
+        && (target_arch == "x86_64" || target_arch == "x86" || target_arch == "aarch64")
+    {
         println!("cargo:rustc-cfg=ptrace_supported");
+    }
+    if target_os == "linux" && (target_arch == "x86_64" || target_arch == "x86") {
+        println!("cargo:rustc-cfg=ptrace_default");
     }
 }